[hopem,r=]
Adds support for changing the min_part_hours value on swift
rings post-install and resyncing rings across cluster on-demand.

Default min-hours is now 0 to allow any number of storage units
to join and enforce a rebalance and resync of rings across the
cluster.

Also refactored code and cleaned up style.