[hopem,r=]
Stop all proxy services if leader changes during a sync. This
will currently require manual intervention to resolve since
we do not yet have leader election in juju, the corosync cluster
leader will need to be switched back to the original leader if
possible since, given that we were mid-sync, that is the only
node with correct rings.