Add statsd logging support to the swift-proxy charm
By default, statsd metrics can be sent by swift-proxy
for diagnostic and monitoring purposes, but are disabled
by default. This change exposes charm config settings
that allow it to be enabled by setting 'statsd_host'
to a non-empty value. 'statsd_port' and 'statsd_sample_rate'
are also supported for changing the destination port
and rate at which metrics are collected.

Closes-Bug: #1729771

Change-Id: I2d5cca233d48022073b5975c06c3da2b1896d8d9