Fix S3 API for >= Kilo
From Kilo onwards the swift-proxy charm is
misconfiguring the swift3 middleware such that
the api is unable to respond to any requests.
We fix this by providing working config for
Kilo onwards.

NOTE: see LP for full explanation but due to
problems with package version mismatches in the
UCA this patch only fixes Trusty Kilo, (L is
EOL) and Xenial Mitaka.

Change-Id: Ice5690e7f06ffc78dd20b53b67dffc6bd72b2613
Closes-Bug: 1738063